feat:add tela de login e cadastro responsivo.

+++ backend/src/controllers/ControllerEmprestimo.ts
@@ -15,7 +15,10 @@ export class EmprestimoController {
         try {
             let emprestimo: CreateEmprestimo = req.body;
             const resposta = await this.emprestimoService.novoEmprestimo(emprestimo);
-            res.status(200).json(resposta);
+            res.status(200).json({
+                dados: resposta,
+                mensagem: "Emprestimo cadastrado com sucesso."
+            });
         } catch (error) {
             next(error)
         }
@@ -25,7 +28,10 @@ export class EmprestimoController {
         try {
             let id = Number(req.params.id)
             const resposta = await this.emprestimoService.listarEmprestimoId(id);
-            res.status(200).json(resposta);
+            res.status(200).json({
+                dados: resposta,
+                mensagem: "Emprestimo encontrado com sucesso."
+            });
         } catch (error) {
             next(error)
         }
@@ -33,8 +39,11 @@ export class EmprestimoController {
 
     public listarEmprestimos = async (req: Request, res: Response, next: NextFunction) => {
         try {
-            let alunos = await this.emprestimoService.listarEmprestimos();
-            res.status(200).json(alunos);
+            const resposta = await this.emprestimoService.listarEmprestimos();
+            res.status(200).json({
+                dados: resposta,
+                mensagem: "Emprestimos listados com sucesso."
+            });
         } catch (error) {
             next(error);
         }
@@ -42,9 +51,15 @@ export class EmprestimoController {
 
     public atualizarEmprestimo = async (req: Request, res: Response, next: NextFunction) => {
         try {
-            let id = Number(req.params.id)
+            const id = Number(req.params.id);
+            if (isNaN(id)) {
+                throw new Error("ID inválido.");
+            }
             const resposta = await this.emprestimoService.atualizarEmprestimo(id, req.body);
-            res.status(200).json(resposta);
+            res.status(200).json({
+                dados: resposta,
+                mensagem: "Emprestimo atualiza com sucesso."
+            });
         } catch (error) {
             next(error)
         }
@@ -52,9 +67,12 @@ export class EmprestimoController {
 
     public delete = async (req: Request, res: Response, next: NextFunction) => {
         try {
-            let id = Number(req.params.id)
-            const resposta = await this.emprestimoService.delete(id);
-            res.status(200).json(resposta);
+            const id = Number(req.params.id);
+            if (isNaN(id)) {
+                throw new Error("ID inválido.");
+            }
+            await this.emprestimoService.delete(id);
+            res.status(204).send();
         } catch (error) {
             next(error)
         }
